Item types to none.
To be doubly sure that we don't pick up something that's already been picked up, we set the item's type to NONE when the item's game object is destroyed, and then check if the item type is NONE when we test the trigger.  Given this worked for powerups, this should probably work here as well.