Use timestamps of the new snapshot messages to detect missing packets
This improves the smoothness of ship movements under bad network
conditions, especially with high packet loss ratios.

This patch changes the data sent in the m_server_timestamp field of
NewPlayerSnapshotToClientMessages to transmit the server's
Time.fixedTime field which matches the fact that the individual
snapshots always contain movement over Time.fixedDeltaTime seconds.

The client uses this timestamp to detect if any packets inbetween
were missing (or even if we  got duplicates or messages out of order,
although that should never happen as long as we use unity's "reliable"
transmission channel).

If missing packets are detected, synthetic messages are calculated
based on the new message and the latest message we received, and
added to the ring buffer, so that the actual per-frame code can
be kept simple and rely on the fact that the ring buffer always
contains packets which are fixedDeltaTime apart in simulation time.

Furthermore, the number of missing packets is also used as basis
for the time sync calculation, as we now exactly know how many
new packets the server actually did sent since the last time we
did the resync.