delay the time resync as much as possible to allow to process more than one new snapshot at once
This addresses the issue that occurs when packets arrive late followed by
several packets received in short succession, as could happen if data
is buffered somewhere on your network path, or by packet loss, when unity's
reliable conncetion fowrds two packets at once after it detected that
one was missing.

Consider the following example:
We do not receive new packets for 70ms (4 physics ticks), and then 4 new
packages within 1ms. On Reception of the first packet, we would detect
that we are out of sync by 3 frames, triggering a hard sync, and during the
processing of the 3 following packets, we would be out of sync by -3
frames again, doing another hard sync inbetween. However, if we had
processed all 4 packets at once and only synced after all of them,
we would have processed all 4 packets worth of 67ms simulation time wich
closely matches the 70ms elapsed time, and no hard resync is neccessary
at all.

This implementation works cby ounting the number of received (and enqueued)
packets for which no time resync was done so far, and delaying the re-sync
until we need the data for rendering a new frame. All packets which arrived
during the rendering of the last frame will be processed at once.