Do not render the UI collision mesh.
This improves the rendering performace. You can expect to observe
a boost in the range of 2% to 8% of your frame rate, depending on
your hardware and your settings.

In my test scenario on my machine, framerate on a particular spot
in vault went from ~160 to ~170 (+6%).

Overload uses a mesh to render the distorted menu and UI elements (also
in game). For the menu to be usable with the mouse, the distortion
has to be inverted. This is done by _another_ mesh which is almost
identical to the render mesh and is used only for CPU side raycasting in
UIManager.FindMousePosition().

However, for dynamically creating this collision only mesh, Luke
Schneider's concept of "UnityRenderLayer" is used, But this has the
effect that we now have an "collision ui layer" lying around which has
its own Renderer and whivh actually renders the collision mesh
_each frame_, for no benefeit at all.

This patch simply disables the renderer for the UI collision layer.