Disable audio playback on the server.
Currently, the dedicated server plays all music and sound effects, just
the volume is forced to 0 (in this patch, there is a commented-out code
bit which overrides this volume, so you can hear all the audio it is
playing).

Disabling the music and sound playback on the server should reduce the
server CPU load a bit.

Note: Unity has an AudioManager which has a "Disable Audio" property.
Ideally, we could find a way to access this, as it also should prevent
the server from trying to grab an audio device in the first place.
However, there is no direct way from Unity C# scripts to access this
functionality. There is a suggestion on the net to do something like that:
https://gist.github.com/benblo/d8b31eab90195208cc38

	var audioManager = AssetDatabase.LoadAllAssetsAtPath("ProjectSettings/AudioManager.asset")[0];
	var serializedManager = new SerializedObject(audioManager);
	var prop = serializedManager.FindProperty("m_DisableAudio");
	prop.boolValue = true;
	serializedManager.ApplyModifiedProperties();

However, we don't have AssetDatabase in our Unity Classes, and I have found
no other way to get access to that asset.