Fix JIP invisible players bug
The issue is that when we do JIP, we completely disable the new player
during the 3 seconds prematch countdown, not only on all clients, but on
the server, too. If a second client connects during that time,
the disabled player will not be sent to the new client at all.

This patch modifies SetReady() so that on the server, we still disable
the player ship and the collider, but we do not completely disable the
Player object itself.