Lag compensation: add option to limit how far ships might dive into walls.
This adds the new menu option "LIMIT SHIPS DIVING INTO WALLS" to the
advanced lag compensation settings. If activated (>0), ship extrapolation
will do an addition `SphereCast` to check wether extrapolation would place
ships into walls (or lava), and limit the maximum dive-in range.

The menu option controls the percentage of the diameter of the ship's
enclosing sphere which must always remain visible, setting it to 100
will prevent ships from diving in at all.

Drawback if this is activated:
- Ships will abruplty stop and appear "stuck" at the wall for short moments
  (instead of diving in further)
- Additional computational overhead if enabled (setting > 0), although
  it did not make a noticable difference in my tests. YMMV

Default is 0, so people have to explicitely enable it if they want it.