JIP: If a non-modded client joins an ongoing match with a custom map,      warn the client that OLMod might be required.
This will delay the JIP operation on this client by a few seconds to
give the user time to read the messages, but it does not prevent
joining altogether if the client actually has the custom map.

OLMod clients aren't delayed in any case, as well as games on standard
maps.

This will also inform non-modded clients when they enter the lobby before
the game is started.

Note that we have no way to know if an unmodded client has a map installed
or not, and as soon we send the scene load command to the client, it will
disconnect itself if it is not found.