Add a controller name remapping feature
This is an attempt to work around an issue on Wormaus' system where
his joystick sometimes appears as "Unknown".

The game will look for the file "controllers.remapmod" in the game
data directory (where the pilot files are stored). Each line can
contain a remap definition of the form

oldName->newName

Lines beginning with `#` are ignored as comments. Make sure to not
add any additional whitespace characeters anywhere in the line. Be
warned that some controller names are reported with spaces at the
beginning and/or end. Look up the exact names in your `<pilot>.xconfig`
file to get the correct spelling. The remappings are case sensitive.