Skip expensive Client.ReconcileServerPlayerState in situations where it is not necessary.
Most of the time, the error is zero or below any noticable threshold. In
such situations, the client-side  player ship resimulation can be skipped
completely, which will make a difference in busy games and especially on
high ping, as the amount of ticks to re-simulate depends on the network
latency between server and client.