Reduce "Player Joining" timeout from 60s to 5s
Sometimes, we see the server hang in the "Player Joining" phase for
a whole minute, where other players are unable to join. Usually, the
"Player Joining" state is shown for fractions of a second, so reducing
this timeout (it is probably a remnant from the original matchmaking
system) to just 5 seconds should be still more than enough.