Explicitly set the order of the transpilers to DrawMpMatchSetup() This patch fixes Multiplayer "Advanced Settings" menu on linux builds.
If a function is patched multiple times, the order the patches are applied
is not well-defined - it depends on the order the patch methids are found
in the assembly, which may differ for different build systems.

The menu was broken before due to multiple transpilers adding new menu
options at the same spot. Currently, there are a total of 10(!)
transpilers to this method in 8 different files. This patch explicitly
applies the order by specifying `[Harmony.Priority]` annotations to each
of them. Note that a higher number means the transpiler is applied
earlier, so I use the `Priority.Normal -n` scheme to denote the n-th
place in the series.

The order explicitely set here matches the order the transpilers are
applied in the windows build, so it has no further effect on the offical
olmod binaries which are distributed.